refactor: use BIGSERIAL for primary keys and centralize test DB setup
- Update Flyway migrations (V1, V2, V4-V11) to use `BIGSERIAL` for auto-incrementing primary keys.
- Create `PostgresContainer` utility to centralize Testcontainers configuration, Flyway migration, and Exposed DB connection logic.
- Refactor `UserRepositoryImplTest` to use the new `PostgresContainer` utility.
- Add initial test class and database setup for `MovieRepositoryImplTest`.

+++ src/main/resources/db/migration/V6__movie_genres_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS movie_genre(
     movie_id BIGINT NOT NULL,
-    genre_id INT NOT NULL,
+    genre_id BIGINT NOT NULL,
 
     PRIMARY KEY (movie_id, genre_id),
     FOREIGN KEY (movie_id) REFERENCES movies(id) ON DELETE CASCADE,
