+++ src/main/resources/db/migration/V8__seats_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS seats(
-    id INT PRIMARY KEY,
+                                    id BIGSERIAL PRIMARY KEY,
     room_id INT NOT NULL,
     row_label TEXT NOT NULL,
     seat_number INT NOT NULL,
